sources/requirements.txt
flask==2.2.5
tensorflow==2.10.0
pillow==9.5.0